requirements.txt
requests==2.32.5
python-dotenv==1.2.2
google==3.0.0
google-genai==1.65.0